// components/AllApplication/type.ts
export interface Application {
  id: string;
  created_at: string;
  updated_at: string;
  deleted_at: string | null;
  is_active: boolean;
  lead_id: number | string;
  company_status_id: number;
  agent_user_id: number;
  company_name: string;
  registration_no: string;
  is_micro_business: boolean;
  number_of_employees: string;
  estimated_turnover: string;
  current_address_line1: string;
  current_address_line2: string;
  current_address_line3: string | null;
  current_address_line4: string | null;
  current_postcode: string;
  owner_partner_name: string;
  owner_partner_dob: string;
  owner_partner_dobstring: string;
  home_address_line1: string;
  home_address_line2: string;
  home_address_line3: string | null;
  home_address_line4: string | null;
  home_postcode: string;
  time_at_current_address_months: number | null;
  previous_address_line1: string | null;
  previous_address_line2: string | null;
  previous_address_line3: string | null;
  previous_address_line4: string | null;
  previous_postcode: string | null;
  time_at_previous_address_months: number | null;
  previous_address2_line1: string | null;
  previous_address2_line2: string | null;
  previous_address2_line3: string | null;
  previous_address2_line4: string | null;
  previous_postcode2: string | null;
  time_at_previous_address2_months: number | null;
  gas_provider: string | null;
  gas_renewal_date: string | null;
  gas_spending_band: string | null;
  electric_provider: string | null;
  electric_renewal_date: string | null;
  electric_spending_band: string | null;
  telecoms_provider: string | null;
  telecoms_renewal_date: string | null;
  telecoms_spending_band: string | null;
  gi_provider: string | null;
  gi_renewal_date: string | null;
  gi_spending_band: string | null;
  partner_user_id: number | null;
  account_manager_user_id: number | null;
  created_user_id: number;
  last_modified_user_id: number;
  deleted_user_id: number | null;
  deleted_datetime: string | null;
  is_deleted: boolean;
  account_manager_user_name: string | null;
  contracts_processed: number;
  contracts_issold: boolean | null;
  contracts_soldleadid: number | null;
  meterstring: string;
  primary_telephone_number: string | null;
  original_lead_source_campaign_string: string | null;
  original_lead_source_campaign_id: number | null;
  is_close_to_renewal: boolean;
  has_overdue_callbacks: boolean;
  has_renewal_callbacks: boolean;
  has_company_callbacks: boolean;
  ced: string | null;
  sold_supplier_name: string | null;
  username: string | null;
  contract_type: string | null;
  created_by: number;
  created_by_name?: string;
  deleted_by: number | null;
  business_type: number;
}